Add names attributes to the scheduled_events 'time' slot
Allow the 'time' column of the events data.frame argument to the
'scheduled_events' method to be either an integer or a Date vector. A
Date vector is coerced to an integer vector and represented as the
number of days since 1970-01-01 and the dates are added as names to
the integer vector slot 'time'.