Modification to tests that use proportions
The model previously would have multiplied the p = 10 or p = -1 by the
number of individuals in the compartment and resulted in moving a
negative number of individuals or a number of individuals greater than
the number in the node. This no longer occurs because the the n is
obtained from the binomial distribution via gsl.

However, errors are still raised in other places of the code if
negative numbers or >1 are supplied to p.